Fix errors when dropping old tables
and add a new flag 'runfresh' to the script in order to
run behat tests on a new fresh database

+++ htdocs/testing/frameworks/behat/classes/util.php
@@ -121,9 +121,10 @@ class BehatTestingUtil extends TestingUtil {
             throw new MaharaBehatTestException('This method can be only used by Behat CLI tool');
         }
 
-        self::reset_dataroot();
-        self::drop_dataroot();
-        self::drop_database(true);
+        if (table_exists(new XMLDBTable('config'))) {
+            self::drop_dataroot();
+            self::drop_database(true);
+        }
     }
 
     /**
